多租户完善

+++ mp-mulitenant/src/main/java/com/dewen/entity/User.java
@@ -11,5 +11,5 @@ public class User {
     @TableId(type = IdType.AUTO)
     private Integer id;
     private String name;
-    private Integer tenantId;
+    private Long tenantId;
 }
